Add directionality argument to grid mixins.
Almost every grid mixin accepts a new argument named `first` for the reading
direction of the document. The exceptions are mixins like `container` or `full`
that are symmetrical. If no value is passed explicitly to the new `first`
arguments, `$document-direction` is passed as the default value. Its default
value is `left` for LTR documents. Override `$document-direction` to create a
RTL grid.