Remove trailing slashes from site URLs
Some of the URLs in the list of sites using Susy had trailing slashes
while others did not. I sanded off the bonus trailing slashes for
consistency and simplicity. Save them bytes.

I left the trailing slash on the Concordia discors link, since removing
it would have resulted in a redirect every time it was followed.